Search of action log and job log

+++ src/main/java/com/web_service/repository/ETLRequestRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.web_service.entity.ETLEntity;
 
-public interface ETLRequestRepository extends JpaRepository<ETLEntity, Integer>{
-	ETLEntity findByRequestId(Integer requestId);
+public interface ETLRequestRepository extends JpaRepository<ETLEntity, Long>{
+	ETLEntity findByRequestId(Long requestId);
 }
